+++ src/commands/ability.mjs
@@ -1,7 +1,7 @@
 import { ActionRowBuilder, SlashCommandBuilder, EmbedBuilder } from "discord.js";
 import fetch from "node-fetch";
 import jsdom from "jsdom";
-import tests from "../modules/tests.mjs";
+import tests from "../modules/nameFinders.mjs";
 import handlers from "../modules/handlers.mjs";
 const { JSDOM } = jsdom;
